backend/requirements.txt
asgiref==3.6.0
Django==3.2.18
django-cors-headers==3.14.0
djangorestframework==3.14.0
pytz==2023.3
sqlparse==0.4.3
